+++ pkg/webui/console/api/index.js
@@ -15,8 +15,6 @@
 import axios from 'axios'
 import TTS, { STACK_COMPONENTS_MAP, AUTHORIZATION_MODES } from 'ttn-lw'
 
-import TTS, { STACK_COMPONENTS_MAP, AUTHORIZATION_MODES } from 'ttn-lw'
-
 import toast from '@ttn-lw/components/toast'
 
 import {